Ave-Faison website create with bootstrap
This is my assignment for the ave website with html, css and bootstrap. It's not much better but it is created using HTML, CSS and boot, not javascript. so I'll add functionalities like an e-commerce store in it after learning js.
Also, it's only ready for mobile and large screens and not ready for tablet devices.